extract SearchSnippet.BuildWindow from both search services (GREEN)
