+++ src/DfE.CheckPerformanceData.Application/ContentBlocks/ContentBlockSearchService.cs
@@ -1,6 +1,6 @@
-using System.Net;
 using DfE.CheckPerformanceData.Application.Common;
 using DfE.CheckPerformanceData.Application.PageTree;
+using DfE.CheckPerformanceData.Application.Search;
 
 namespace DfE.CheckPerformanceData.Application.ContentBlocks;
 
@@ -81,30 +81,7 @@ public sealed class ContentBlockSearchService(
         ["/guidance"] = "CYPMD help and support",
     };
 
-    // Produces safe snippet HTML: everything is HTML-encoded, then the matched term is
-    // wrapped in a <mark>. The only markup that can reach the page is that <mark>.
-    private static string BuildSnippet(string? plainText, string term)
-    {
-        var plain = plainText ?? string.Empty;
-        var idx = plain.IndexOf(term, StringComparison.OrdinalIgnoreCase);
-
-        if (idx < 0)
-        {
-            var head = plain.Length <= 200 ? plain : plain[..200] + "…";
-            return WebUtility.HtmlEncode(head);
-        }
-
-        var start = Math.Max(0, idx - 60);
-        var length = Math.Min(plain.Length - start, 200);
-        var window = plain.Substring(start, length);
-        var wi = window.IndexOf(term, StringComparison.OrdinalIgnoreCase);
-
-        var before = WebUtility.HtmlEncode(window[..wi]);
-        var match = WebUtility.HtmlEncode(window.Substring(wi, term.Length));
-        var after = WebUtility.HtmlEncode(window[(wi + term.Length)..]);
-
-        var prefix = start > 0 ? "…" : string.Empty;
-        var suffix = start + length < plain.Length ? "…" : string.Empty;
-        return $"{prefix}{before}<mark>{match}</mark>{after}{suffix}";
-    }
+    // Windowed HTML-encode + <mark>-wrap logic lives on the shared SearchSnippet helper.
+    private static string BuildSnippet(string? plainText, string term) =>
+        SearchSnippet.BuildWindow(plainText ?? string.Empty, term);
 }
